Refactored nav logic into NavWrapper
